+++ uniffi_bindgen/src/bindings/python/templates/ObjectTemplate.py
@@ -76,22 +76,22 @@ class {{ impl_name }}:
 
 class {{ ffi_converter_name }}:
     {%- if obj.is_trait_interface() %}
-    _handle_map = ConcurrentHandleMap()
+    _handle_map = UniffiHandleMap()
     {%- endif %}
 
     @staticmethod
-    def lift(value: int):
+    def lift(value: UniffiHandle):
         return {{ impl_name }}._make_instance_(value)
 
     @staticmethod
-    def lower(value: {{ protocol_name }}):
+    def lower(value: {{ type_name }}):
         {%- match obj.imp() %}
         {%- when ObjectImpl::Struct %}
         if not isinstance(value, {{ impl_name }}):
             raise TypeError("Expected {{ impl_name }} instance, {} found".format(type(value).__name__))
         return value._uniffi_handle
         {%- when ObjectImpl::Trait %}
-        return {{ ffi_converter_name }}._handle_map.insert(value)
+        return {{ ffi_converter_name }}._handle_map.new_handle(value)
         {%- endmatch %}
 
     @classmethod
@@ -100,5 +100,5 @@ class {{ ffi_converter_name }}:
         return cls.lift(ptr)
 
     @classmethod
-    def write(cls, value: {{ protocol_name }}, buf: _UniffiRustBuffer):
+    def write(cls, value: {{ type_name }}, buf: _UniffiRustBuffer):
         buf.write_u64(cls.lower(value))