+++ uniffi_core/src/ffi/rustfuture.rs
@@ -105,7 +105,7 @@ pub enum RustFuturePoll {
 ///
 /// The Rust side of things calls this when the foreign side should call [rust_future_poll] again
 /// to continue progress on the future.
-pub type RustFutureContinuationCallback = extern "C" fn(callback_data: *const (), RustFuturePoll);
+pub type RustFutureContinuationCallback = extern "C" fn(callback_data: Handle, RustFuturePoll);
 
 // === Public FFI API ===
 
@@ -141,7 +141,7 @@ where
 pub unsafe fn rust_future_poll<ReturnType, UT>(
     handle: Handle,
     callback: RustFutureContinuationCallback,
-    data: *const (),
+    data: Handle,
 ) where
     dyn RustFutureFfi<ReturnType>: HandleAlloc<UT>,
 {
@@ -212,7 +212,6 @@ derive_ffi_traits!(impl<UT> HandleAlloc<UT> for dyn RustFutureFfi<f64>);
 derive_ffi_traits!(impl<UT> HandleAlloc<UT> for dyn RustFutureFfi<Handle>);
 derive_ffi_traits!(impl<UT> HandleAlloc<UT> for dyn RustFutureFfi<crate::RustBuffer>);
 derive_ffi_traits!(impl<UT> HandleAlloc<UT> for dyn RustFutureFfi<()>);
-derive_ffi_traits!(impl<UT> HandleAlloc<UT> for dyn RustFutureFfi<*const ::std::ffi::c_void>);
 
 /// Thread-safe storage for [RustFutureContinuationCallback] data
 ///
@@ -230,8 +229,8 @@ enum ContinuationDataCell {
     /// The future has been cancelled, any future `store` calls should immediately result in the
     /// continuation being called with `RustFuturePoll::Ready`.
     Cancelled,
-    /// Continuation set, the next time `wake()`  is called is called, we should invoke it.
-    Set(RustFutureContinuationCallback, *const ()),
+    /// Continuation set, the next time `wake()` is called, we should invoke it.
+    Set(RustFutureContinuationCallback, Handle),
 }
 
 impl ContinuationDataCell {
@@ -241,7 +240,7 @@ impl ContinuationDataCell {
 
     /// Store new continuation data if we are in the `Empty` state.  If we are in the `Waked` or
     /// `Cancelled` state, call the continuation immediately with the data.
-    fn store(&mut self, callback: RustFutureContinuationCallback, data: *const ()) {
+    fn store(&mut self, callback: RustFutureContinuationCallback, data: Handle) {
         match self {
             Self::Empty => *self = Self::Set(callback, data),
             Self::Set(old_callback, old_data) => {
@@ -288,11 +287,6 @@ impl ContinuationDataCell {
         matches!(self, Self::Cancelled)
     }
 }
-
-// ContinuationDataCell is Send + Sync as long we handle the *const () pointer correctly
-
-unsafe impl Send for ContinuationDataCell {}
-unsafe impl Sync for ContinuationDataCell {}
 
 /// Wraps the actual future we're polling
 struct WrappedFuture<F, T, UT>
@@ -438,7 +432,7 @@ where
         })
     }
 
-    fn poll(self: Arc<Self>, callback: RustFutureContinuationCallback, data: *const ()) {
+    fn poll(self: Arc<Self>, callback: RustFutureContinuationCallback, data: Handle) {
         let ready = self.is_cancelled() || {
             let mut locked = self.future.lock().unwrap();
             let waker: std::task::Waker = Arc::clone(&self).into();
@@ -503,8 +497,8 @@ where
 /// x86-64 machine . By parametrizing on `T::ReturnType` we can instead monomorphize by hand and
 /// only create those functions for each of the 13 possible FFI return types.
 #[doc(hidden)]
-pub trait RustFutureFfi<ReturnType> {
-    fn ffi_poll(self: Arc<Self>, callback: RustFutureContinuationCallback, data: *const ());
+pub trait RustFutureFfi<ReturnType>: Send + Sync {
+    fn ffi_poll(self: Arc<Self>, callback: RustFutureContinuationCallback, data: Handle);
     fn ffi_cancel(&self);
     fn ffi_complete(&self, call_status: &mut RustCallStatus) -> ReturnType;
     fn ffi_free(self: Arc<Self>);
@@ -517,7 +511,7 @@ where
     T: LowerReturn<UT> + Send + 'static,
     UT: Send + 'static,
 {
-    fn ffi_poll(self: Arc<Self>, callback: RustFutureContinuationCallback, data: *const ()) {
+    fn ffi_poll(self: Arc<Self>, callback: RustFutureContinuationCallback, data: Handle) {
         self.poll(callback, data)
     }
 
@@ -601,13 +595,15 @@ mod tests {
     /// Poll a Rust future and get an OnceCell that's set when the continuation is called
     fn poll(rust_future: &Arc<dyn RustFutureFfi<RustBuffer>>) -> Arc<OnceCell<RustFuturePoll>> {
         let cell = Arc::new(OnceCell::new());
-        let cell_ptr = Arc::into_raw(cell.clone()) as *const ();
-        rust_future.clone().ffi_poll(poll_continuation, cell_ptr);
+        let handle = <OnceCell<RustFuturePoll> as HandleAlloc<crate::UniFfiTag>>::new_handle(
+            Arc::clone(&cell),
+        );
+        rust_future.clone().ffi_poll(poll_continuation, handle);
         cell
     }
 
-    extern "C" fn poll_continuation(data: *const (), code: RustFuturePoll) {
-        let cell = unsafe { Arc::from_raw(data as *const OnceCell<RustFuturePoll>) };
+    extern "C" fn poll_continuation(data: Handle, code: RustFuturePoll) {
+        let cell = <OnceCell<RustFuturePoll> as HandleAlloc<crate::UniFfiTag>>::get_arc(data);
         cell.set(code).expect("Error setting OnceCell");
     }